requirements.txt
Django==2.0.4
ExifRead==2.1.2
gphoto2==1.8.2
pyserial==3.4
pytz==2018.4
PyYAML==3.12
inotify==0.2.9
redis==2.10.6
